+++ tests/population/test_population_exports.py
@@ -8,15 +8,61 @@ import variopt.algorithms.population.csa as csa_algorithms
 import variopt.algorithms.population.de as de_algorithms
 import variopt.algorithms.population.ga as ga_algorithms
 import variopt.algorithms.population.generational_ga as generational_ga_algorithms
+import variopt.algorithms.population.generational_ga.state as generational_ga_state
 import variopt.algorithms.population.permutation as permutation_algorithms
 import variopt.algorithms.population.restricted_tournament_ga as restricted_tournament_ga_algorithms
 import variopt.algorithms.population.species_ga as species_ga_algorithms
+
+EXPECTED_POPULATION_ALL = (
+    "ClearingGAProfile",
+    "ClearingGeneticAlgorithmOptimizer",
+    "CSAOptimizer",
+    "CSAProfile",
+    "DEProfile",
+    "DifferentialEvolutionOptimizer",
+    "GAProfile",
+    "GeneticAlgorithmOptimizer",
+    "GenerationalGAOptimizerState",
+    "GenerationalGAPopulationMember",
+    "GenerationalGAVariant",
+    "InversionMutation",
+    "OrderCrossover",
+    "RestrictedTournamentGAProfile",
+    "RestrictedTournamentGeneticAlgorithmOptimizer",
+    "SpeciesConservingGeneticAlgorithmOptimizer",
+    "SpeciesGAProfile",
+    "SwapMutation",
+)
+
+EXPECTED_ROOT_ALGORITHMS_ALL = (
+    "AlgorithmProfile",
+    "ClearingGeneticAlgorithmOptimizer",
+    "ClearingGAProfile",
+    "DEProfile",
+    "DifferentialEvolutionOptimizer",
+    "GAProfile",
+    "GeneticAlgorithmOptimizer",
+    "ScipyMinimizeKernel",
+    "ScipyMinimizeMethod",
+    "RestrictedTournamentGAProfile",
+    "RestrictedTournamentGeneticAlgorithmOptimizer",
+    "SpeciesConservingGeneticAlgorithmOptimizer",
+    "SpeciesGAProfile",
+    "StructuredHillClimbKernel",
+    "StructuredIteratedLocalSearchKernel",
+    "StructuredKickPolicy",
+    "StructuredStochasticNeighborhoodKernel",
+    "StructuredScheduledLocalSearchKernel",
+    "StructuredVariableNeighborhoodKernel",
+    "StructuredVariableNeighborhoodStage",
+)
 
 
 class PopulationFacadeExportTests:
     """Lock the canonical population-family facade surface."""
 
     def test_population_facade_reexports_population_family_entry_points(self) -> None:
+        assert tuple(population_algorithms.__all__) == EXPECTED_POPULATION_ALL
         assert population_algorithms.CSAOptimizer is csa_algorithms.CSAOptimizer
         assert population_algorithms.CSAProfile is csa_algorithms.CSAProfile
         assert population_algorithms.DEProfile is de_algorithms.DEProfile
@@ -32,12 +78,13 @@ class PopulationFacadeExportTests:
         assert population_algorithms.ClearingGeneticAlgorithmOptimizer is clearing_ga_algorithms.ClearingGeneticAlgorithmOptimizer
         assert population_algorithms.RestrictedTournamentGAProfile is restricted_tournament_ga_algorithms.RestrictedTournamentGAProfile
         assert population_algorithms.RestrictedTournamentGeneticAlgorithmOptimizer is restricted_tournament_ga_algorithms.RestrictedTournamentGeneticAlgorithmOptimizer
+        assert population_algorithms.GenerationalGAOptimizerState is generational_ga_state.GenerationalGAOptimizerState
+        assert population_algorithms.GenerationalGAPopulationMember is generational_ga_state.GenerationalGAPopulationMember
+        assert population_algorithms.GenerationalGAVariant is generational_ga_state.GenerationalGAVariant
 
         internal_names = (
             "GENERATIONAL_GA_EXECUTION_MODELS",
             "GenerationalGAGenerationCommit",
-            "GenerationalGAOptimizerState",
-            "GenerationalGAPopulationMember",
             "ask_generational_ga",
             "create_initial_generational_ga_state",
             "sort_generational_ga_population",
@@ -48,6 +95,7 @@ class PopulationFacadeExportTests:
         assert all(not hasattr(population_algorithms, name) for name in internal_names)
 
     def test_root_algorithms_facade_remains_convenience_reexport(self) -> None:
+        assert tuple(root_algorithms.__all__) == EXPECTED_ROOT_ALGORITHMS_ALL
         assert root_algorithms.DEProfile is population_algorithms.DEProfile
         assert root_algorithms.DifferentialEvolutionOptimizer is population_algorithms.DifferentialEvolutionOptimizer
         assert root_algorithms.GAProfile is population_algorithms.GAProfile